Handle attributes with values over 30 characters specially
Mastodon will not accept such values for displayname or mail (we already
limited nickname to this length). We cannot arbitrarily shorten them
(chop off someone's name or part of an email address?).

For displayname we send the nickname instead which we know is shorter
than 30 chars, and the user can change this after account creation.

For mail we just do not send it and let Mastodon deal with it.